[Move]EnchantUtilを作成し、移動
